fix bug sending email, bug viGuard API, bug email saldo, update sending notification
